+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Looking for changes that affect our C API? See the [C API Changelog](lib/c-api/C
 
 ### Changed
 - #2864 wasmer-cli: remove wasi-experimental-io-devices from default builds
+### Changed
+- #2868 Removed loupe crate dependency
 
 ### Fixed
 - [#2829](https://github.com/wasmerio/wasmer/pull/2829) Improve error message oriented from JS object.